requirements.txt
tqdm
elsapy
beautifulsoup4
requests
pymupdf
psycopg2-binary
